Move drag-and-drop components to their own files
